backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.12
langchain-groq==0.1.3
groq==0.5.0
asyncpg==0.29.0
pydantic==2.6.3
python-dotenv==1.0.1